+++ static/booking.css
@@ -110,6 +110,7 @@ main {
     margin-bottom: 20px;
 }
 
+
 hr {
     background-color: #E8E8E8;
     border: none;
@@ -137,7 +138,6 @@ hr {
 .attraction-img {
     width: 250px;
     height: 200px;
-
     margin: 0 30px 0px 0px;
 }
 
@@ -196,7 +196,7 @@ hr {
     margin-bottom: 15px;
 }
 
-.information-item > input {
+.information-item > input, .tpfield {
     font-size: 16px;
     font-weight: 500;
     color: #000000;
@@ -204,14 +204,19 @@ hr {
     border: 1px solid #E8E8E8;
     border-radius: 5px;
     width:200px;
-    height:38px
+    height:38px;
+    outline:none;
 }
 
-
-.information-item > input::placeholder {
+.information-item > input::placeholder, .tpfield::placeholder {
     color: #757575;
     font-weight: 500;
     font-size: 16px;
+}
+
+.information-item {
+    display: flex;
+    align-items: center;
 }
 
 .attention {
