+++ static/thankyou.css
@@ -1,0 +1,210 @@
+* {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+    font-family: Noto Sans TC, sans-serif;
+}
+
+.navbar-container {
+    width: 100%;
+    height: 54px;
+    align-items: center;
+    justify-content: center;
+    display: flex;
+    z-index: 99;
+    position: fixed;
+    top: 0;
+    background-color: #ffffff;
+}
+
+.navbar {
+    max-width: 1200px;
+    width: 62.5%;
+    height: 34px;
+    top: 10px;
+    display: flex;
+    justify-content: space-between;
+}
+
+.navbar > h2 {
+    font-size: 30px;
+    font-weight: 700;
+    line-height: 34px;
+    color: #448899;
+    width: 150px;
+    height: 34px;
+    margin: 0;
+    cursor: pointer;
+}
+
+.navbar-lists{
+    width: auto;
+    display: flex;
+}
+
+.btn-arrow-left{
+    padding: 9px 0px 9px 15px;
+}
+
+.btn-arrow-right{
+    padding: 9px 15px 9px 0px;
+}
+
+
+.navbar-list {
+    cursor: pointer;
+    background: none;
+    border: none;
+    font-size: 16px;
+    font-weight: 500;
+    color: #666666;
+    height: 34px;
+    padding: 10px;
+    width: auto;
+    margin: 0;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+hr {
+    background-color: #E8E8E8;
+    border: none;
+    width: 100vw;
+    height: 1px;
+}
+
+#hr-top{
+    position: fixed;
+    margin-top: 0;
+    z-index: 98;
+}
+
+main {
+    width: 100%;
+    height: auto;
+    padding: 40px 0px 40px 0px;
+    margin: auto;
+    margin-top: 54px;
+}
+
+.information-title {
+    height: 27px;
+    display: flex;
+    align-items: center;
+    font-size: 19px;
+    font-weight: 700;
+    line-height: 16px;
+    color: #666666;
+}
+
+.information-title-marginbtn20 {
+    margin-bottom: 20px;
+}
+
+.information-container {
+    display: flex;
+}
+
+.image-container {
+    width: 300px;
+    height: auto;
+    margin: 0 30px 0px 0px;
+}
+
+.subtitle {
+    height: 24px;
+    margin-bottom: 10px;
+    font-size: 16px;
+    font-weight: 700;
+    color: #666666;
+}
+
+
+img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+.information-item {
+    display: flex;
+    align-items: center;
+}
+
+.attention {
+    font-size: 16px;
+    font-weight: 700;
+    color: #666666;
+}
+
+footer {
+    width: 100%;
+    height: 104px;
+    padding: 40px 0;  
+    background: #757575;
+    text-align: center;
+    display: flex;
+}
+
+.footer-text {
+    height: 24px;
+    margin: auto;
+    font-size: 16px;
+    font-weight: 700;
+    color: #ffffff;
+}
+
+.order-information {
+    font-weight: 500;
+}
+
+@media (min-width: 1200px) {
+    .navbar, main, .information-hr {
+        width: 1200px;
+    }
+}
+
+@media (max-width: 1200px) {
+    .navbar, main{
+        width: 90%;
+    }
+    .information-hr {
+        width: 100%;
+    }
+
+}
+
+@media (max-width: 600px) {
+    .navbar, main{
+        width: 100%;
+    }
+
+    .navbar{
+        margin-left: 10px;
+        margin-right: 10px;
+    }
+
+    .information-container{
+        display: block;
+    }
+    .information-title, .information-container{
+        margin-left: 10px;
+    }
+
+    .information-title {
+        margin-bottom: 30px;
+    }
+
+    .image-container {
+        width: calc(100% - 10px);
+        margin-bottom: 30px;
+    }
+
+}
+
+@media (max-width: 360px) {
+    header, main, footer, hr {
+        width: 360px;
+    }
+}
